% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merged_snps.R
\name{current_snp}
\alias{current_snp}
\title{Converts a SNP id to its current SNP id.}
\usage{
current_snp(snps, mapping = FALSE, keep.orig.snp.if.not.merged = TRUE)
}
\arguments{
\item{snps}{A numeric vector of SNP reference ids (without the 'rs').}

\item{mapping}{If FALSE (default), return a vector of current ids; if TRUE, return a data frame with two columns,}

\item{keep.orig.snp.if.not.merged}{If \code{TRUE} (default), the returned vector of current SNP ids will contain original SNP ids for those SNPs that were not found in the translation table, otherwise, \code{NA}s will be returned instead.}
}
\value{
Either a vector of current SNP ids or a data frame of original SNPs and current SNPs.
}
\description{
refSNP(rs) evolve in time and may be merged to earlier reference ids. This function allows to resolve a SNP id to its current id. This function uses information provided in columns 1 and 7, i.e., rsHigh and rsCurrent, respectively, in file RsMergeArch.bcp.gz for \emph{Homo sapiens}, from NCBI's dbSNP, build 149, Genome Reference Consortium 38, patch 7. Source data was originally downloaded from dbSNP ftp: \href{ftp://ftp.ncbi.nih.gov/snp/organisms/human_9606_b149_GRCh38p7/database/organism_data/RsMergeArch.bcp.gz}{RsMergeArch.bcp.gz}.
}
