% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merged_snps.R
\name{replace_na_with_snps}
\alias{replace_na_with_snps}
\title{Replace NAs with SNP ids}
\usage{
replace_na_with_snps(old_snps, new_snps)
}
\value{
A vector of SNP ids.
}
\description{
Replace \code{NA} in \code{new_snps} with SNP ids from \code{old_snps}.
Replacement is done position-wise.
}
