% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merged_snps.R
\name{snp_translation_table}
\alias{snp_translation_table}
\title{Translation table of merged SNPs}
\usage{
snp_translation_table()
}
\value{
A data frame. First column is rsHigh (\code{old_snp}) and rsCurrent
(\code{upd_snp}).
}
\description{
Returns the translation table of merged SNPs for \emph{Homo sapiens}, from
NCBI's dbSNP, build 149, Genome Reference Consortium 38, patch 7. This
function uses information provided in columns 1 and 7, i.e., rsHigh and
rsCurrent, respectively, in the file \href{ftp://ftp.ncbi.nih.gov/snp/organisms/human_9606_b149_GRCh38p7/database/organism_data/RsMergeArch.bcp.gz}{RsMergeArch.bcp.gz}. For more information check
\href{https://www.ncbi.nlm.nih.gov/SNP/snp_db_table_description.cgi?t=RsMergeArch}{dbSNP Column Description for table: RsMergeArch}.
}
